requirements.txt
# Web Crawling and Content Extraction
scrapy>=2.12.0  # Python 3.13 compatible
trafilatura>=1.6.2
htmldate>=1.5.2
requests>=2.31.0
requests-ratelimiter>=0.4.1

# Database
psycopg2-binary>=2.9.10  # Python 3.13 compatible
sqlalchemy>=2.0.23
alembic>=1.12.1
redis>=5.0.1

# AI APIs
anthropic>=0.8.1
openai>=1.6.1

# Configuration Management
pydantic>=2.9.0  # Python 3.13 compatible
pydantic-settings>=2.5.0
python-dotenv>=1.0.0
PyYAML>=6.0

# Notifications
slack-sdk==3.26.1

# Utilities
python-json-logger==2.0.7

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-mock>=3.12.0

# Development tools (optional)
black==23.12.1
flake8==7.0.0
mypy==1.7.1
